refactored code is now functional
