+++ frontend/App.js
@@ -1,5 +1,5 @@
-export default class App{
-  function toggleCatSearch(){
+class App {
+   toggleCatSearch(){
     let catForm = document.getElementById('cat-search');
     if(catForm.style.display === 'none'){
       catForm.style.display = 'block';
@@ -9,7 +9,7 @@ export default class App{
     }
   }
 
-  function toggleResults(){
+   toggleResults(){
     let results = document.getElementById('search-results');
     if(results.style.display === 'none') {
       results.style.display = 'block';
@@ -19,7 +19,8 @@ export default class App{
     }
   }
 
-  function toggleNewExp(){
+   toggleNewExp(){
+    console.log('clicked!');
     let newExp = document.getElementById('exp-form');
     if(newExp.style.display === 'none'){
       newExp.style.display = 'block';
@@ -29,7 +30,7 @@ export default class App{
     }
   }
 
-  function toggleNewCat(){
+   toggleNewCat(){
     let newCat = document.getElementById('cat-form');
     if(newCat.style.display === 'none'){
       newCat.style.display = 'block';
@@ -39,4 +40,82 @@ export default class App{
     }
   }
 
+   removeAllChildNodes(parent) {
+      while (parent.firstChild) {
+          parent.removeChild(parent.firstChild);
+      }
+  }
+
+   populateGrid(grid, results){
+    let elements  = results
+    let gridName = `${grid}-grid`
+    let targetGrid = document.getElementById(gridName);
+    app.removeAllChildNodes(targetGrid);
+
+    //make a new tile for the exp and then append it to the new-grid
+  for(let i = 0; i<elements.length; i++) {
+      let container = document.createElement('div');
+      container.className = 'experience--container'
+
+      let photo = document.createElement('img');
+      photo.href = elements[i]['photo'];
+      photo.className = 'exp--photo';
+
+      let title = document.createElement('h3');
+      title.textContent = elements[i]['title'];
+      title.className = 'exp--title';
+
+      let description = document.createElement('h2');
+      description.textContent = elements[i]['description'];
+      description.className = 'exp--description';
+
+      let date = document.createElement('h3');
+      date.textContent = elements[i]['date'];
+      date.className = 'exp--date';
+
+      let location =  document.createElement('h3');
+      location.textContent = elements[i]['location'];
+      location.classNameName = 'exp--location';
+
+      let coordinates = document.createElement('h3');
+      coordinates.textContent = elements[i]['coordinates'];
+      coordinates.className = 'exp--coordinates';
+
+      container.appendChild(photo);
+      container.appendChild(title);
+      container.appendChild(description);
+      container.appendChild(date);
+      container.appendChild(location);
+      container.appendChild(coordinates);
+
+      targetGrid.appendChild(container);
+    };
+  }
+
+   populateCats(results){
+    let elements  = results
+    let targetGrid = document.getElementById('cat-options');
+    app.removeAllChildNodes(targetGrid);
+
+    //make a new tile for the exp and then append it to the new-grid
+  for(let i = 0; i<elements.length; i++) {
+      let option = document.createElement('option');
+      option.value = elements[i].name;
+      option.textContent = elements[i].name;
+
+      targetGrid.appendChild(option);
+    };
+  }
+
+   getButtons(){
+    //JS variables for all nav & search buttons
+    return {
+      refresh: document.getElementById('refresh'),
+      newExp: document.getElementById('new-exp'),
+      catSearchBtn: document.getElementById('cat-search-btn'),
+      newExpSubmit: document.getElementById('new-exp-submit'),
+      newCatSubmit: document.getElementById('new-cat-submit'),
+      newCat: document.getElementById('new-cat')
+    };
+  }
 }
