+++ frontend/index.html
@@ -10,7 +10,7 @@
     <meta name="description" content="A site built for positivity">
     <meta name="author" content="Ray Ockenfels">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="index.js"></script>
+
 </head>
 
 <body>
@@ -83,4 +83,7 @@
     <div id='experience-grid'>
     </div>
   </section>
+  <script src="App.js"></script>
+  <script src="Adaptor.js"></script>
+  <script src="index.js"></script>
 </body>
